doc_warehouse/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pytesseract
pdf2image
sentence-transformers
faiss-cpu
openai
dotenv